Force light theme for Gingerbread and below
Dialogs in Android don't always behave as expected when it comes
to theming. When running on Gingerbread, the country picker may
use black background while also drawing the country names in
black text color, making them difficult to read.

The solutions, as noted on e.g.
https://groups.google.com/forum/?fromgroups=#!topic/actionbarsherlock/lGuxn7YjTY0
is to call setInverseBackgroundForced(true).